Colocate operation specs and flatten transcription modules
