+++ scinoephile/llms/delineation/operation.py
@@ -1,11 +1,13 @@
 #  Copyright 2017-2026 Karl T Debiec. All rights reserved. This software may be modified
 #  and distributed under the terms of the BSD license. See the LICENSE file for details.
-"""Language-agnostic transcription delineation configuration."""
+"""Delineation operation configuration."""
 
 from __future__ import annotations
 
 from scinoephile.core.llms import OperationSpec
-from scinoephile.llms.delineation import DelineationManager, DelineationPrompt
+
+from .manager import DelineationManager
+from .prompt import DelineationPrompt
 
 __all__ = ["DELINEATION_OPERATION_SPEC"]
 
@@ -16,4 +18,4 @@ DELINEATION_OPERATION_SPEC = OperationSpec(
     manager_cls=DelineationManager,
     prompt_cls=DelineationPrompt,
 )
-"""Operation specification for transcription delineation."""
+"""Operation specification for delineation."""
