+++ scinoephile/llms/punctuation/operation.py
@@ -1,11 +1,13 @@
 #  Copyright 2017-2026 Karl T Debiec. All rights reserved. This software may be modified
 #  and distributed under the terms of the BSD license. See the LICENSE file for details.
-"""Language-agnostic transcription punctuation configuration."""
+"""Punctuation operation configuration."""
 
 from __future__ import annotations
 
 from scinoephile.core.llms import OperationSpec
-from scinoephile.llms.punctuation import PunctuationManager, PunctuationPrompt
+
+from .manager import PunctuationManager
+from .prompt import PunctuationPrompt
 
 __all__ = ["PUNCTUATION_OPERATION_SPEC"]
 
@@ -17,4 +19,4 @@ PUNCTUATION_OPERATION_SPEC = OperationSpec(
     prompt_cls=PunctuationPrompt,
     list_fields={f"query.{PunctuationPrompt.src_1}": 10},
 )
-"""Operation specification for transcription punctuation."""
+"""Operation specification for punctuation."""
