--- scinoephile/multilang/transcription/__init__.py
@@ -1,7 +1,0 @@
-#  Copyright 2017-2026 Karl T Debiec. All rights reserved. This software may be modified
-#  and distributed under the terms of the BSD license. See the LICENSE file for details.
-"""Code for transcription operations involving multiple languages.
-
-Package hierarchy (modules may import from any above):
-* delineation / punctuation
-"""
