+++ scinoephile/multilang/yue_zho/transcription/delineation.py
@@ -1,6 +1,6 @@
 #  Copyright 2017-2026 Karl T Debiec. All rights reserved. This software may be modified
 #  and distributed under the terms of the BSD license. See the LICENSE file for details.
-"""Text for LLM correspondence for simplified written Cantonese delineation."""
+"""Written Cantonese/standard Chinese transcription delineation."""
 
 from __future__ import annotations
