--- scinoephile/multilang/yue_zho/transcription/delineation/__init__.py
@@ -1,15 +1,0 @@
-#  Copyright 2017-2026 Karl T Debiec. All rights reserved. This software may be modified
-#  and distributed under the terms of the BSD license. See the LICENSE file for details.
-"""Code related to written Cantonese/standard Chinese transcription delineation."""
-
-from __future__ import annotations
-
-from .prompt import (
-    YueDelineationVsZhoPromptYueHans,
-    YueDelineationVsZhoPromptYueHant,
-)
-
-__all__ = [
-    "YueDelineationVsZhoPromptYueHans",
-    "YueDelineationVsZhoPromptYueHant",
-]
