--- scinoephile/multilang/yue_zho/transcription/punctuation/__init__.py
@@ -1,22 +1,0 @@
-#  Copyright 2017-2026 Karl T Debiec. All rights reserved. This software may be modified
-#  and distributed under the terms of the BSD license. See the LICENSE file for details.
-"""Code related to written Cantonese/standard Chinese transcription punctuation.
-
-Package hierarchy (modules may import from any above):
-* prompt
-* manager
-"""
-
-from __future__ import annotations
-
-from .manager import YueZhoPunctuationManager
-from .prompt import (
-    YuePunctuationVsZhoPromptYueHans,
-    YuePunctuationVsZhoPromptYueHant,
-)
-
-__all__ = [
-    "YuePunctuationVsZhoPromptYueHans",
-    "YuePunctuationVsZhoPromptYueHant",
-    "YueZhoPunctuationManager",
-]
